+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 Date format: DD/MM/YYYY
 
+## [2.0.0] - Theming Update - [26/02/2021]
+
+- 🔶 `ToastTheme` now works properly
+- 🌌 `StyledToast` now uses Snackbar's Theme Data
+- 📔 Improved snackbar fidelity in example app and cookbook
+- ⏳ Change animation controller duration when updated on the widget
+- **NEW** 🦡 `Toast.semanticsLabel`
+- **NEW** ⏰ `Toast.reverseAnimationDuration`
+
 ## [1.0.1] - The major update - [21/02/2021]
 
 - 📦 Example app
@@ -7,11 +16,18 @@ Date format: DD/MM/YYYY
 - 📂 Formatted files
 - 📃 Improved documentation
 - 🍪 New Cookbook
+- **NEW** 📜 `showPermanentToast`
 
 ## [1.0.0+1] - [20/02/2021]
 
 - Readme update
 
-## [1.0.0] - [20/02/2021]
+## [1.0.0] - Initial Release - [20/02/2021]
 
-- Initial release
+- Show toasts:
+  - `showToast`
+  - `showMultiToast`
+  - `showStyledToast`
+  - `showAndroidToast`
+  - `showPlatformToast`
+  - `showSequencialToast`